Support win-auth not being standard auth
Attempt to fix edge case of win-auth not being the standard authentication. 
Currently you can add prefix to host and authenticate, but cant deploy extensions.
Alternatively you can add prefix to prefix variable and not be able to authenticate. (but would probably be able to deploy)

This change allows prefix to work for win-auth authentication if needed.